mgr/cephadm: move _refresh_hosts_and_daemons to serve.py
(cherry picked from commit d73dc54e4ee9150e750afe779546d0242ad176c2)
