+++ src/pybind/mgr/cephadm/module.py
@@ -1384,75 +1384,6 @@ To check that the host is reachable:
         self.log.debug(f'Refreshed OSDSpec previews for host <{host}>')
         return True
 
-    def _refresh_hosts_and_daemons(self) -> None:
-        bad_hosts = []
-        failures = []
-
-        @forall_hosts
-        def refresh(host):
-            if self.cache.host_needs_check(host):
-                r = self._check_host(host)
-                if r is not None:
-                    bad_hosts.append(r)
-            if self.cache.host_needs_daemon_refresh(host):
-                self.log.debug('refreshing %s daemons' % host)
-                r = self._refresh_host_daemons(host)
-                if r:
-                    failures.append(r)
-
-            if self.cache.host_needs_registry_login(host) and self.registry_url:
-                self.log.debug(f"Logging `{host}` into custom registry")
-                r = self._registry_login(host, self.registry_url,
-                                         self.registry_username, self.registry_password)
-                if r:
-                    bad_hosts.append(r)
-
-            if self.cache.host_needs_device_refresh(host):
-                self.log.debug('refreshing %s devices' % host)
-                r = self._refresh_host_devices(host)
-                if r:
-                    failures.append(r)
-
-            if self.cache.host_needs_osdspec_preview_refresh(host):
-                self.log.debug(f"refreshing OSDSpec previews for {host}")
-                r = self._refresh_host_osdspec_previews(host)
-                if r:
-                    failures.append(r)
-
-            if self.cache.host_needs_new_etc_ceph_ceph_conf(host):
-                self.log.debug(f"deploying new /etc/ceph/ceph.conf on `{host}`")
-                r = self._deploy_etc_ceph_ceph_conf(host)
-                if r:
-                    bad_hosts.append(r)
-
-        refresh(self.cache.get_hosts())
-
-        health_changed = False
-        if 'CEPHADM_HOST_CHECK_FAILED' in self.health_checks:
-            del self.health_checks['CEPHADM_HOST_CHECK_FAILED']
-            health_changed = True
-        if bad_hosts:
-            self.health_checks['CEPHADM_HOST_CHECK_FAILED'] = {
-                'severity': 'warning',
-                'summary': '%d hosts fail cephadm check' % len(bad_hosts),
-                'count': len(bad_hosts),
-                'detail': bad_hosts,
-            }
-            health_changed = True
-        if failures:
-            self.health_checks['CEPHADM_REFRESH_FAILED'] = {
-                'severity': 'warning',
-                'summary': 'failed to probe daemons or devices',
-                'count': len(failures),
-                'detail': failures,
-            }
-            health_changed = True
-        elif 'CEPHADM_REFRESH_FAILED' in self.health_checks:
-            del self.health_checks['CEPHADM_REFRESH_FAILED']
-            health_changed = True
-        if health_changed:
-            self.set_health_checks(self.health_checks)
-
     def _refresh_host_daemons(self, host) -> Optional[str]:
         try:
             out, err, code = self._run_cephadm(
