+++ src/pybind/mgr/cephadm/tests/fixtures.py
@@ -5,6 +5,7 @@ from contextlib import contextmanager
 
 from ceph.deployment.service_spec import PlacementSpec, ServiceSpec
 from cephadm.module import CEPH_DATEFMT
+from cephadm.serve import CephadmServe
 
 try:
     from typing import Any, Iterator, List
@@ -118,7 +119,7 @@ def with_host(m: CephadmOrchestrator, name, refresh_hosts=True):
     # type: (CephadmOrchestrator, str) -> None
     wait(m, m.add_host(HostSpec(hostname=name)))
     if refresh_hosts:
-        m._refresh_hosts_and_daemons()
+        CephadmServe(m)._refresh_hosts_and_daemons()
     yield
     wait(m, m.remove_host(name))
