+++ src/pybind/mgr/cephadm/tests/test_upgrade.py
@@ -6,6 +6,7 @@ import pytest
 from ceph.deployment.service_spec import ServiceSpec
 from cephadm import CephadmOrchestrator
 from cephadm.upgrade import CephadmUpgrade
+from cephadm.serve import CephadmServe
 from .fixtures import _run_cephadm, wait, cephadm_module, with_host, with_service
 
 
@@ -73,7 +74,7 @@ def test_upgrade_run(use_repo_digest, cephadm_module: CephadmOrchestrator):
                     )
                 ])
             )):
-                cephadm_module._refresh_hosts_and_daemons()
+                CephadmServe(cephadm_module)._refresh_hosts_and_daemons()
 
             cephadm_module.upgrade._do_upgrade()
